+++ src/app/farming/actions.ts
@@ -12,7 +12,7 @@ import {
   isD1Store,
 } from 'app/inventory/stores-helpers';
 import { supplies } from 'app/search/d1-known-values';
-import { refresh } from 'app/shell/refresh';
+import { refresh } from 'app/shell/refresh-events';
 import { ThunkResult } from 'app/store/types';
 import { CancelToken, withCancel } from 'app/utils/cancel';
 import { infoLog } from 'app/utils/log';
