+++ src/app/inventory/drag-events.ts
@@ -1,0 +1,3 @@
+import { Observable } from 'app/utils/observable';
+
+export const isDragging$ = new Observable<boolean>(false);
