+++ src/app/vendors/SingleVendor.tsx
@@ -24,7 +24,7 @@ import {
 } from '../inventory/selectors';
 import { DimStore } from '../inventory/store-types';
 import { loadingTracker } from '../shell/loading-tracker';
-import { refresh$ } from '../shell/refresh';
+import { refresh$ } from '../shell/refresh-events';
 import { loadAllVendors } from './actions';
 import { toVendor } from './d2-vendors';
 import type { VendorsState } from './reducer';
