+++ src/app/vendors/Vendors.tsx
@@ -33,7 +33,7 @@ import {
 } from '../inventory/selectors';
 import { DimStore } from '../inventory/store-types';
 import { loadingTracker } from '../shell/loading-tracker';
-import { refresh$ } from '../shell/refresh';
+import { refresh$ } from '../shell/refresh-events';
 import { loadAllVendors } from './actions';
 import {
   D2VendorGroup,
